drm: sun4i: bypass RGB interface dotclock check when panel is used
The RGB interface code of sun4i-drm has a dotclock check, which is used
to prevent not supported resolutions when a bridge is used.

However, it prevents most of the panels in panel-simple driver from
being used, as the dotclock may not be divided precisely to the dotclock
in panel-simple driver.

Bypass the dotclock check when a panel is used.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>