ARM: dts: suniv: add initial DTSI file for suniv and F1C100s
As we have the support for suniv pin controller and CCU now, add a
initial DTSI for it. All currently known SoCs with suniv should have the
same peripheral set and share most of the content of the suniv DTSI.

F1C100s is one product with the suniv die, which has a 32MiB co-packaged
DDR1 DRAM chip. As there's nothing special for it, add a dummy DTSI file
for it.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>